% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratestat.R
\name{ratestat}
\alias{ratestat}
\title{Statistics on singing rate}
\usage{
ratestat(folder, nlb)
}
\arguments{
\item{folder}{Folder containing the labels.}

\item{nlb}{Maximum level of file label to analyze.}
}
\description{
Get an estimate of the mean singing rate of a bird and the uncertainty of the estimate.
}
\keyword{rate}
\keyword{singing}
