% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rec_lottery.R
\name{rec_lottery}
\alias{rec_lottery}
\title{Pick sound recordings at random}
\usage{
rec_lottery(x, y, n, fs)
}
\arguments{
\item{x}{Input folder}

\item{y}{Output folder}

\item{n}{Total number of files to draw}

\item{fs}{Sections for file sorting}
}
\description{
Pick sound recordings at random
}
